requirements.txt
faiss-cpu
sentence-transformers
torch
transformers
numpy
pandas
scikit-learn
PyPDF2
python-docx
reportlab
markdown
streamlit
tqdm
pyyaml
argparse